+++ src/rs/iterators/mod.rs
@@ -1,0 +1,12 @@
+/* SPDX-FileCopyrightText: © 2026 Decompollaborate */
+/* SPDX-License-Identifier: MIT */
+
+mod iter_mapfile_syms_by_name;
+mod iter_mapfile_syms_by_vram;
+mod iter_mapfile_syms_by_vrom;
+mod iter_segment_syms_by_name;
+
+pub use iter_mapfile_syms_by_name::IterMapFileSymsByName;
+pub use iter_mapfile_syms_by_vram::IterMapFileSymsByVram;
+pub use iter_mapfile_syms_by_vrom::IterMapFileSymsByVrom;
+pub use iter_segment_syms_by_name::IterSegmentSymsByName;